% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{log_hooks_disable}
\alias{log_hooks_disable}
\title{Disable automatic data I/O logging hooks}
\usage{
log_hooks_disable()
}
\value{
Invisibly \code{NULL}.
}
\description{
Restores all functions patched by \code{\link[=log_hooks_enable]{log_hooks_enable()}} to their originals.
Safe to call even if hooks were never enabled.
}
\examples{
\dontrun{
log_hooks_disable()
}

}
\seealso{
\code{\link[=log_hooks_enable]{log_hooks_enable()}}, \code{\link[=with_log]{with_log()}}
}
